--- src/main/java/com/bae/persistence/rest/GameController.java
@@ -1,5 +1,0 @@
-package com.bae.persistence.rest;
-
-public class GameController {
-
-}
