+++ src/main/java/com/bae/rest/rest/GameController.java
@@ -1,0 +1,5 @@
+package com.bae.rest.rest;
+
+public class GameController {
+
+}
